backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-filter==2.4.0
django-imagekit==4.0.2
djangorestframework==3.12.2
pilkit==2.0
Pillow==8.1.1
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
